+++ packages/patternfly-4/react-core/src/components/Pagination/index.d.ts
@@ -1,1 +1,0 @@
-export { default as Pagination, PaginationVariant, PerPageOptions, PaginationTitles } from './Pagination';