+++ packages/patternfly-4/react-integration/CHANGELOG.md
@@ -3,6 +3,33 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+# [2.2.0](https://github.com/patternfly/patternfly-react/compare/@patternfly/react-integration@2.1.2...@patternfly/react-integration@2.2.0) (2019-07-02)
+
+
+### Features
+
+* **Pagination:** Pagination Component Typescript Conversion ([#2256](https://github.com/patternfly/patternfly-react/issues/2256)) ([8a29a20](https://github.com/patternfly/patternfly-react/commit/8a29a20))
+
+
+
+
+
+## [2.1.2](https://github.com/patternfly/patternfly-react/compare/@patternfly/react-integration@2.1.1...@patternfly/react-integration@2.1.2) (2019-07-01)
+
+**Note:** Version bump only for package @patternfly/react-integration
+
+
+
+
+
+## [2.1.1](https://github.com/patternfly/patternfly-react/compare/@patternfly/react-integration@2.1.0...@patternfly/react-integration@2.1.1) (2019-07-01)
+
+**Note:** Version bump only for package @patternfly/react-integration
+
+
+
+
+
 # [2.1.0](https://github.com/patternfly/patternfly-react/compare/@patternfly/react-integration@2.0.0...@patternfly/react-integration@2.1.0) (2019-07-01)